requirements.txt
Flask==3.0.0
flask-cors==4.0.0
Werkzeug==3.0.1
torch>=2.2.0
torchvision>=0.17.0
Pillow>=10.0.0
numpy>=1.24.0
